apps/windows-sender/requirements.txt
cryptography>=44.0.0
Pillow>=11.0.0
requests>=2.32.0
